Add Circe codecs for GeoTiffInfo and related classes (#3128)
* Move GeoTiffInfo to geotrellis.raster.io.geotiff.reader

Moving out of GeoTiffReader class. This is a often used class in client code and it makes little sense to tuck it away in a reader.

* Add circe codecs for GeoTiffInfo and all related classes

Encoding GeoTiffInfo can be useful for caching purposes. To minimize the implicit resolution time we're tucking away codecs into companion objects and making heavy use of macro encoder/decoder derivation where we can.

* Merge TiffTagsReader with TiffTags

Move to minimze weirdness in code base.
- we don't need another class to hold a function
- make private methods private to minmize public API surface
- unroll the super weird method extension for tag reading
- apply methods should not do reading
