+++ docs/CHANGELOG.rst
@@ -19,6 +19,8 @@ API Changes & Project structure changes
   - **Change:** ``geotrellis.raster.summary.polygonal.[Multi]TilePolygonalSummaryHandler`` replaced with ``geotrellis.raster.summary.polygonal.PolygonalSummary. Users should expect to implement concrete subclasses of ``geotrellis.raster.summary.GridVisitor`` and pass those to the new polygonalSummary methods. There are a number of default GridVisitor implementations provided for simple operations in `geotrellis.raster.summary.visitors``
   - **New:** Created the ``geotrellis.raster.mapalgebra.focal.ZFactor`` to improve slope calculations (`#3014 <https://github.com/locationtech/geotrellis/pull/3014>`_).
   - **Change:** ``Grid.dimensions`` is now ``Dimensions[N]`` instead of ``(N, N)`` (`#3124 <https://github.com/locationtech/geotrellis/pull/3124> _)
+  - **Change:** ``GeoTiffInfo`` moved out of ``GeoTiffReader`` into ``geotrellis.raster.io.geotiff.reader`` (`#3128 <https://github.com/locationtech/geotrellis/pull/3128> _)
+  - **New:** Added cerci encoders and decoders for ``GeoTiffInfo`` and realted classes (`#3128 <https://github.com/locationtech/geotrellis/pull/3128> _)
 
 - ``geotrellis.layer``
