+++ raster/src/main/scala/geotrellis/raster/io/geotiff/StorageMethod.scala
@@ -16,7 +16,9 @@
 
 package geotrellis.raster.io.geotiff
 
-import geotrellis.raster._
+import io.circe._
+import io.circe.syntax._
+import cats.syntax.either._
 
 abstract sealed class StorageMethod extends Serializable
 
@@ -50,4 +52,35 @@ object Striped extends StripedStorageMethod {
   def apply(): Striped = new Striped(None)
 
   implicit def objectToStorageMethod(s: StripedStorageMethod): Striped = Striped()
+}
+
+object StorageMethod {
+  implicit val storageMethodDecoder: Decoder[StorageMethod] =
+    new Decoder[StorageMethod] {
+      final def apply(c: HCursor): Decoder.Result[StorageMethod] = {
+        c.downField("storageType").as[String].flatMap {
+          case "striped" =>
+            val rowsPerStrip = c.downField("rowsPerStrip").as[Option[Int]].getOrElse(None)
+            Right(new Striped(rowsPerStrip))
+          case _ =>
+            (c.downField("cols").as[Int], c.downField("rows").as[Int]) match {
+              case (Right(cols), Right(rows)) => Right(Tiled(cols, rows))
+              case _ =>
+                Left(DecodingFailure(s"No cols / rows were specified for the Tiled storage method.", c.history))
+            }
+        }
+      }
+    }
+
+  implicit val storageMethodEncoder: Encoder[StorageMethod] =
+    new Encoder[StorageMethod] {
+      final def apply(a: StorageMethod): Json = a match {
+        case _: Striped => Json.obj(("storageType", "striped".asJson))
+        case Tiled(cols, rows) => Json.obj(
+          ("storageType", "tiled".asJson),
+          ("cols", cols.asJson),
+          ("rows", rows.asJson)
+        )
+      }
+    }
 }
