+++ raster/src/main/scala/geotrellis/raster/io/geotiff/compression/Compression.scala
@@ -16,6 +16,36 @@
 
 package geotrellis.raster.io.geotiff.compression
 
+import io.circe._
+import io.circe.syntax._
+import cats.syntax.either._
+
 trait Compression extends Serializable {
   def createCompressor(segmentCount: Int): Compressor
+}
+
+object Compression {
+  implicit val compressionDecoder: Decoder[Compression] =
+    new Decoder[Compression] {
+      final def apply(c: HCursor): Decoder.Result[Compression] = {
+        c.downField("compressionType").as[String].map {
+          case "NoCompression" => NoCompression
+          case _ =>
+            c.downField("level").as[Int] match {
+              case Left(_)  => DeflateCompression()
+              case Right(i) => DeflateCompression(i)
+            }
+        }
+      }
+    }
+
+  implicit val compressionEncoder: Encoder[Compression] =
+    new Encoder[Compression] {
+      final def apply(a: Compression): Json = a match {
+        case NoCompression =>
+          Json.obj(("compressionType", "NoCompression".asJson))
+        case d: DeflateCompression =>
+          Json.obj(("compressionType", "Deflate".asJson), ("level", d.level.asJson))
+      }
+    }
 }
