+++ raster/src/main/scala/geotrellis/raster/io/geotiff/tags/DataSampleFormatTags.scala
@@ -17,6 +17,8 @@
 package geotrellis.raster.io.geotiff.tags
 
 import monocle.macros.Lenses
+import io.circe._
+import io.circe.generic.semiauto._
 
 @Lenses("_")
 case class DataSampleFormatTags(
@@ -24,3 +26,8 @@ case class DataSampleFormatTags(
   maxSampleValue: Option[Array[Long]] = None,
   minSampleValue: Option[Array[Long]] = None
 )
+
+object DataSampleFormatTags {
+  implicit val dataSampleFormatTagsDecoder: Decoder[DataSampleFormatTags] = deriveDecoder[DataSampleFormatTags]
+  implicit val dataSampleFormatTagsEncoder: Encoder[DataSampleFormatTags] = deriveEncoder[DataSampleFormatTags]
+}
