+++ raster/src/main/scala/geotrellis/raster/io/geotiff/tags/GeoTiffTags.scala
@@ -17,6 +17,8 @@
 package geotrellis.raster.io.geotiff.tags
 
 import monocle.macros.Lenses
+import io.circe._
+import io.circe.generic.semiauto._
 
 @Lenses("_")
 case class GeoTiffTags(
@@ -29,3 +31,8 @@ case class GeoTiffTags(
   metadata: Option[String] = None,
   gdalInternalNoData: Option[Double] = None
 )
+
+object GeoTiffTags {
+  implicit val geotiffTagsDecoder: Decoder[GeoTiffTags] = deriveDecoder[GeoTiffTags]
+  implicit val geotiffTagsEncoder: Encoder[GeoTiffTags] = deriveEncoder[GeoTiffTags]
+}
