+++ raster/src/main/scala/geotrellis/raster/io/geotiff/tags/NonBasicTags.scala
@@ -17,6 +17,8 @@
 package geotrellis.raster.io.geotiff.tags
 
 import monocle.macros.Lenses
+import io.circe._
+import io.circe.generic.semiauto._
 
 @Lenses("_")
 case class NonBasicTags(
@@ -38,3 +40,8 @@ case class NonBasicTags(
   halftoneHints: Option[Array[Int]] = None,
   predictor: Option[Int] = None
 )
+
+object NonBasicTags {
+  implicit val nonBasicTagsEncoder: Encoder[NonBasicTags] = deriveEncoder[NonBasicTags]
+  implicit val nonBasicTagsDecoder: Decoder[NonBasicTags] = deriveDecoder[NonBasicTags]
+}
